Create the platform handler system.
Platform-specific actions will be performed through the IPlatformHandler
and IPlatformCallbackHandler interfaces.

The PlatformFactory will pick the IPlatformHandler for the instant platform
and make it available through a singleton.

IPlatformHandlers are always created with reflection (using a no-arg constructor),
so it's safe to have them use platform-specific stuff.

If you add a platform, be sure to add its class name to the list in the PlatformFactory.
Make sure it's no-arg constructor throws an exception if you're NOT on the platform you
intend to handle!