Add installation helper files to the repo
The Info.plist file is, along with a stub executable supplied by Apple,
the main thing you add to the bundle to turn jxm.jar into a full fledged
executable.

jxm.nsi is an NSIS 2.0 installer script file. Run that through NSIS with
the jar and a few other files and out pops the Windows installer.