Sigh.
If you plug in an unactivated unit having previously used an activated one,
the surfer must be made aware of the previous channel list, otherwise it won't
know to send delete channel events.

Speaking of which, delete channel events happen with channel numbers, not SIDs.

And the channel list may be rebuilt with the radio officially off, thanks to the
generous 250 msec wait in the radio commander. So rebuilding the favorite list
needs to know that.